refactor: move schema into one folder
